Specialize Util.writeShortest() for Uint{16,32}
This adds type safety to the picture, so that we end up doing
minimal correct checks on the value supplied. It also allows us
to ditch superfluous use of ByteBufWriteUtils.

+++ bgp/extensions/flowspec/src/main/java/org/opendaylight/protocol/bgp/flowspec/FSIpv6FlowLabelHandler.java
@@ -33,7 +33,7 @@ public final class FSIpv6FlowLabelHandler implements FlowspecTypeParser, Flowspe
         for (final Iterator<FlowLabel> it = list.iterator(); it.hasNext(); ) {
             final FlowLabel label = it.next();
             final ByteBuf protoBuf = Unpooled.buffer();
-            Util.writeShortest(label.getValue().intValue(), protoBuf);
+            Util.writeShortest(label.getValue(), protoBuf);
             NumericOneByteOperandParser.INSTANCE.serialize(label.getOp(), protoBuf.readableBytes(),
                     !it.hasNext(), nlriByteBuf);
             nlriByteBuf.writeBytes(protoBuf);
